requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
python-multipart==0.0.6
flask==2.0.1
requests>=2.27.0
werkzeug==2.0.1

# Database
pymongo==4.6.0
motor==3.3.1

# Document Processing
pdfminer.six==20221105
python-docx==1.0.1

# NLP & ML
spacy==3.7.2
scikit-learn==1.3.2
gensim==4.3.2
transformers==4.35.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.1+cpu
nltk==3.8.1

# Data Processing
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1

# Utilities
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
joblib==1.3.2

# Testing
pytest==7.4.3
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0 